+++ NEWS.md
@@ -65,6 +65,7 @@ sudo make install
 * Fix Mobile Apps deletion bug ([CartoDB/support#2218](https://github.com/CartoDB/cartodb/pull/15135))
 * Update Dataservices API client default version to `0.27.0` (#15134)
 * Allow users to login from forbidden map/dataset page. ([CartoDB/support#2031](https://github.com/CartoDB/support/issues/2031))
+* Fix password validation for SAML ([#15147](https://github.com/CartoDB/cartodb/pull/15147))
 
 4.29.0 (2019-07-15)
 -------------------